+++ tests/get_pred_and_data.py
@@ -1,13 +1,11 @@
 import os
 from typing import Literal
 
-import astropy.units as u
 import numpy as np
 import pandas as pd
 
 import fuellib as fl
 from fuellib._data_locator import get_fueldata_props_dir
-from fuellib.units import convert_temperature
 
 FUELDATA_PROPS_DIR = get_fueldata_props_dir()
 
@@ -17,7 +15,7 @@ def get_pred_and_data(
     prop_name: Literal[
         "Density", "VaporPressure", "Viscosity", "SurfaceTension", "ThermalConductivity"
     ],
-) -> tuple[u.Quantity, u.Quantity, u.Quantity]:
+) -> tuple[fl.units.Quantity, fl.units.Quantity, fl.units.Quantity]:
     # Get the fuel properties based on the GCM
     fuel = fl.fuel(fuel_name)
 
@@ -34,11 +32,11 @@ def get_pred_and_data(
 
     # Filter out invalid (NaN) entries and add units to values
     valid_idxs = ~np.isnan(prop_data)
-    temp_data = convert_temperature(u.Quantity(temp_data[valid_idxs], temp_unit), "K")
-    prop_data = u.Quantity(prop_data[valid_idxs], prop_unit)
+    temp_data = fl.units.Quantity(temp_data[valid_idxs], temp_unit).to("K")
+    prop_data = fl.units.Quantity(prop_data[valid_idxs], prop_unit)
 
     # Compile predictions for the given property
-    pred = u.Quantity(np.zeros(len(temp_data)), prop_unit)
+    pred = fl.units.Quantity(np.zeros(len(temp_data)), prop_unit)
     for i, T in enumerate(temp_data):
         Y_li = fuel.Y_0